DchoBeaconHelp/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
bleak
